requirements.txt
h5py
hydra-colorlog
scikit-learn
torchvision

# add mltools requirements below since we do not want to install mltools in the
# container to keep the flexibility of having it as a git submodule
dotmap==1.3.30
hydra-core==1.3.2
lightning==2.1.2
matplotlib==3.8.2
normflows==1.7.3
numpy==1.26.2
omegaconf==2.3.0
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
PyYAML==6.0.1
rich==13.7.0
scikit_learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
setuptools==68.0.0
torch==2.2.0
torchdiffeq==0.2.3
tqdm==4.65.0
wandb==0.16.0
